Move bookmark state to widget model
